# src/main/resources/swagger/api.yml
openapi: 3.0.3
info:
  title: Jasper KMS
  description: REST API for Jasper Knowledge Management Server
  version: '1.1.0'
  license:
    name: MIT License
    url: 'https://github.com/cjmalloy/jasper/blob/master/LICENSE'
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/cjmalloy/Jasper/1.1.0
paths:
...
  /api/v1/tags:
    post:
      tags:
        - Tagging
      operationId: createTags
      parameters:
        - name: tag
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/tag'
        - name: url
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/url'
        - name: origin
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/origin'
      responses:
        '201':
          description: Created
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags:
        - Tagging
      operationId: deleteTags
      parameters:
        - name: tag
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/tag'
        - name: url
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/url'
        - name: origin
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/origin'
      responses:
        '204':
          description: No Content
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    patch:
      tags:
        - Tagging
      operationId: patchTags
      parameters:
        - name: tags
          in: query
          required: true
          schema:
            type: array
            items:
              $ref: '#/components/schemas/tag'
        - name: url
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/url'
        - name: origin
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/origin'
      responses:
        '204':
          description: No Content
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /api/v1/template:
    get:
      tags:
        - Template
      operationId: getTemplate
      parameters:
        - name: tag
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/template-qtag'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Template'
        '304':
          description: Not Modified
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Template
      operationId: updateTemplate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Template'
        required: true
      responses:
        '204':
          description: No Content
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags:
        - Template
      operationId: createTemplate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Template'
        required: true
      responses:
        '201':
          description: Created
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      tags:
        - Template
      operationId: deleteTemplate
      parameters:
        - name: tag
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/template-qtag'
      responses:
        '204':
          description: No Content
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/template/page:
    get:
      tags:
        - Template
      operationId: getTemplatePage
      parameters:
        - name: page
          in: query
          description: Zero-based page index (0..N)
          required: false
          schema:
            minimum: 0
            type: integer
            default: 0
        - name: size
          in: query
          description: The size of the page to be returned
          required: false
          schema:
            minimum: 1
            type: integer
            default: 10
        - name: sort
          in: query
          description: 'Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.'
          required: false
          schema:
            type: array
            items:
              type: string
            default:
              - 'tag,ASC'
        - name: query
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/query'
        - name: modifiedAfter
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageTemplate'
  /api/v1/template/list:
    get:
      tags:
        - Template
      operationId: getTemplateList
      parameters:
        - name: tags
          in: query
          required: true
          schema:
            maxItems: 100
            minItems: 0
            type: array
            items:
              $ref: '#/components/schemas/template-qtag'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Template'
...
  /api/v1/repl/template:
    get:
      tags:
        - Repl
      operationId: template
      parameters:
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 500
        - name: query
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/query'
        - name: modifiedAfter
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Template'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  schemas:
    name:
      type: string
      description: Display name
      example: Name
      maxLength: 512
    url:
      type: string
      format: uri
      pattern: '([^:/?#]+):(\/\/([^\/?#]*))?([^?#]*)(\?([^#]*))?(#(.*))?'
      description: URL
      example: http://jasper-kms.info/
      maxLength: 4096
...
    origin:
      type: string
      pattern: '@[a-z]+(\.[a-z])*'
      default: ''
      description: Origin
      example: '@origin'
      maxLength: 64
...
    tag:
      type: string
      pattern: '[_+]?[a-z]+(/[a-z]+)*'
      description: Tag
      example: 'tag'
      maxLength: 64
...
    query:
      type: string
      description: Tag Query
      pattern: '(!?([_+]?[a-z]+(\/[a-z]+)*|([_+]?[a-z]+(\/[a-z]+)*)?(@[a-z]+(\.[a-z])*|@\*))|\(!?([_+]?[a-z]+(\/[a-z]+)*|([_+]?[a-z]+(\/[a-z]+)*)?(@[a-z]+(\.[a-z])*|@\*))([ |]!?([_+]?[a-z]+(\/[a-z]+)*|([_+]?[a-z]+(\/[a-z]+)*)?(@[a-z]+(\.[a-z])*|@\*)))*\))([ |:&](!?([_+]?[a-z]+(\/[a-z]+)*|([_+]?[a-z]+(\/[a-z]+)*)?(@[a-z]+(\.[a-z])*|@\*))|\(!?([_+]?[a-z]+(\/[a-z]+)*|([_+]?[a-z]+(\/[a-z]+)*)?(@[a-z]+(\.[a-z])*|@\*))([ |]!?([_+]?[a-z]+(\/[a-z]+)*|([_+]?[a-z]+(\/[a-z]+)*)?(@[a-z]+(\.[a-z])*|@\*)))*\)))*'
      example: '(politics|science):!funny'
      maxLength: 512
    template-tag:
      type: string
      pattern: '(_?[a-z]+(/[a-z]+)*)?'
      description: Template Tag
      example: 'kanban'
      maxLength: 64
    template-qtag:
      type: string
      pattern: '(_?[a-z]+(/[a-z]+)*)?(@[a-z]+(\.[a-z])*)?'
      description: Optionally Qualified Template Tag
      example: 'kanban@origin'
      maxLength: 127
...
    Template:
      type: object
      properties:
        tag:
          $ref: '#/components/schemas/template-tag'
        origin:
          $ref: '#/components/schemas/origin'
        name:
          $ref: '#/components/schemas/name'
        config:
          type: object
        defaults:
          type: object
        schema:
          type: object
        modified:
          type: string
          format: date-time
...
    Sort:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...
    PageTemplate:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/Template'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/Sort'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        'application/problem+json':
          schema:
            $ref: 'https://opensource.zalando.com/problem/schema.yaml#/Problem'
    NotFound:
      description: Not Found
      content:
        'application/problem+json':
          schema:
            $ref: 'https://opensource.zalando.com/problem/schema.yaml#/Problem'
...
    Conflict:
      description: Conflict
      content:
        'application/problem+json':
          schema:
            $ref: 'https://opensource.zalando.com/problem/schema.yaml#/Problem'
    Forbidden:
      description: Forbidden
      content:
        'application/problem+json':
          schema:
            $ref: 'https://opensource.zalando.com/problem/schema.yaml#/Problem'